Remove single page prefix from components
We no longer have a 2-page donation form so this
prefix is now redundant.

Ticket: https://phabricator.wikimedia.org/T378075
