+++ src/components/pages/donation_form/PaymentSummaryContent/PaymentSummaryAnonymous.vue
