+++ tests/unit/components/pages/donation_form/FormSections/PersonalDataSectionDonationReceipt.spec.ts
@@ -1,17 +1,13 @@
 import { mount, VueWrapper } from '@vue/test-utils';
-
-import PersonalDataSection from '@src/components/pages/donation_form/singlePageFormSections/PersonalDataSection.vue';
 import { createStore } from '@src/store/donation_store';
 import { action } from '@src/store/util';
-import { AddressTypeModel } from '@src/view_models/AddressTypeModel';
 import { createFeatureToggle } from '@src/util/createFeatureToggle';
 import { Store } from 'vuex';
 import { TrackingData } from '@src/view_models/TrackingData';
 import { CampaignValues } from '@src/view_models/CampaignValues';
 import { AddressValidation } from '@src/view_models/Validation';
-import { nextTick } from 'vue';
-import AddressTypeBasic from '@src/components/pages/donation_form/AddressTypeBasic.vue';
 import { Salutation } from '@src/view_models/Salutation';
+import PersonalDataSectionDonationReceipt from '@src/components/pages/donation_form/FormSections/PersonalDataSectionDonationReceipt.vue';
 import { FakeBankValidationResource } from '@test/unit/TestDoubles/FakeBankValidationResource';
 
 const testCountry = {
@@ -34,20 +30,20 @@ const salutations: Salutation[] = [
 	},
 ];
 
-describe( 'PersonalDataSection.vue (With Street Autocomplete)', () => {
+describe( 'PersonalDataSectionDonationReceipt.vue', () => {
 	const getWrapper = ( store: Store<any> = createStore() ): { wrapper: VueWrapper<any>, store: Store<any> } => {
-		const wrapper = mount( PersonalDataSection, {
+		const wrapper = mount( PersonalDataSectionDonationReceipt, {
 			props: {
 				assetsPath: '',
 				validateAddressUrl: '',
 				validateEmailUrl: '',
-				validateBankDataUrl: '',
-				validateLegacyBankDataUrl: '',
+				validateBankDataUrl: 'https://localhost:8082',
+				validateLegacyBankDataUrl: 'https://localhost:8082',
 				countries: [ testCountry ],
 				salutations,
 				trackingData: {} as TrackingData,
 				campaignValues: {} as CampaignValues,
-				addressValidationPatterns: { postcode: '', country: null } as AddressValidation,
+				addressValidationPatterns: { postcode: '' } as AddressValidation,
 			},
 			global: {
 				plugins: [ store ],
@@ -58,10 +54,7 @@ describe( 'PersonalDataSection.vue (With Street Autocomplete)', () => {
 					bankValidationResource: new FakeBankValidationResource(),
 				},
 				components: {
-					FeatureToggle: createFeatureToggle( [
-						'campaigns.address_type_steps.preselect',
-						'campaigns.address_field_order.new_order',
-					] ),
+					FeatureToggle: createFeatureToggle( [ 'campaigns.address_type_steps.preselect' ] ),
 				},
 			},
 		} );
@@ -69,19 +62,7 @@ describe( 'PersonalDataSection.vue (With Street Autocomplete)', () => {
 		return { wrapper, store };
 	};
 
-	it( 'sets address type in store when it receives address-type event', () => {
-		const { wrapper, store } = getWrapper();
-
-		store.dispatch = jest.fn();
-		const expectedAction = action( 'address', 'setAddressType' );
-		const expectedPayload = AddressTypeModel.ANON;
-
-		wrapper.findComponent( AddressTypeBasic ).vm.$emit( 'address-type', AddressTypeModel.ANON );
-
-		expect( store.dispatch ).toBeCalledWith( expectedAction, expectedPayload );
-	} );
-
-	it( 'scrolls to top when the donor clicks the previous button', async () => {
+	it( 'scrolls to payment section when button for changing payment data is clicked', async () => {
 		const scrollElement = { scrollIntoView: jest.fn() };
 		Object.defineProperty( document, 'getElementById', { writable: true, configurable: true, value: () => scrollElement } );
 
@@ -92,21 +73,11 @@ describe( 'PersonalDataSection.vue (With Street Autocomplete)', () => {
 		expect( scrollElement.scrollIntoView ).toHaveBeenCalledWith( { behavior: 'smooth' } );
 	} );
 
-	it( 'updates full selected', async () => {
-		const { wrapper } = getWrapper();
-
-		wrapper.findComponent( AddressTypeBasic ).vm.$emit( 'address-type', AddressTypeModel.PERSON );
-		wrapper.findComponent( AddressTypeBasic ).vm.$emit( 'set-full-selected' );
-		await nextTick();
-
-		expect( wrapper.find( '.address-type-person' ).exists() ).toBeTruthy();
-	} );
-
 	it( 'validates the payment section input on page submit', async () => {
 		const { wrapper, store } = getWrapper();
 		store.dispatch = jest.fn().mockResolvedValue( true );
 
-		await wrapper.find( '#submit-btn' ).trigger( 'click' );
+		await wrapper.find( '#donation-form' ).trigger( 'submit' );
 
 		expect( store.dispatch ).toHaveBeenCalledWith( action( 'payment', 'markEmptyValuesAsInvalid' ) );
 	} );
